requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
cryptg==0.4.0
frozenlist==1.4.1
hachoir==3.3.0
idna==3.6
multidict==6.0.5
pillow==10.2.0
pyaes==1.6.1
pyasn1==0.6.0
python-dotenv==1.0.1
rsa==4.9
Telethon==1.34.0
yarl==1.9.4
